systemhotelarski_front/requirements.txt
Django==1.9
django-braces==1.8.1
psycopg2==2.6.1
six==1.10.0
